Create 2 main methods to download data from API

+++ app/src/main/java/com/example/firstkotlinproject/model/Genre.kt
@@ -1,3 +1,3 @@
 package com.example.firstkotlinproject.model
 
-data class Genre(val id: Int, val name: String)
+data class Genre(val id: Long, val name: String)
